requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Vector DB
chromadb>=0.4.0
redis>=5.0.0

# Data Sources
elasticsearch>=8.0.0
gql>=3.5.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
